+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [sparkles] Add sender config to SparklesConfig
 - [sparkles-core] **[WIP]** aarch64 timestamps support
 - [encoder format] Add ticks_per_sec packet type
+- [sparkles-parser] Add small offset when several events recorded with the same timestamp
 
 ## [0.1.3] - 2024-09-23
 - [sparkles-core] New: Add configuration structures: `SparklesConfig` and `LocalStorageConfig`